Refactor login hook and component
loginData state from the useLogin hook now lives in the App component so
that it's available to every page.

The Login component lives in a new components/ directory.

The useLogin hook lives in a new hooks/ directory.
