+++ pages/index.js
@@ -1,65 +1,12 @@
 import Head from 'next/head';
 import getConfig from 'next/config';
 import PropTypes from 'prop-types';
-import { useEffect, useState } from 'react';
 
 import styles from '../styles/Home.module.css';
 
-export const apiStates = {
-  LOADING: 'LOADING',
-  SUCCESS: 'SUCCESS',
-  ERROR: 'ERROR',
-};
+import { Login } from '../components/login';
 
-const useLogin = (url) => {
-  const [data, setData] = useState({
-    state: apiStates.LOADING,
-    error: '',
-    data: [],
-  });
-
-  // Short cut for updating just a single key in the state
-  const setPartialData = (partialData) => setData({ ...data, ...partialData });
-
-  /* eslint-disable react-hooks/exhaustive-deps */
-  useEffect(() => {
-    setPartialData({
-      state: apiStates.LOADING,
-    });
-    fetch(url)
-      .then((response) => {
-        if (response.status >= 200 && response.status <= 299) {
-          return response.json();
-        }
-        throw Error(response.statusText);
-      })
-      .then((jsonData) => {
-        setPartialData({ state: apiStates.SUCCESS, data: jsonData });
-      })
-      .catch(() => {
-        setPartialData({ state: apiStates.ERROR, error: 'fetch failed' });
-      });
-  }, []);
-  /* eslint-enable react-hooks/exhaustive-deps */
-
-  return data;
-};
-
-const Login = ({ loginData }) => (
-  <p>
-    {loginData.state === apiStates.SUCCESS
-      ? `Hello ${loginData.data.username}`
-      : 'Login'}
-  </p>
-);
-
-Login.propTypes = {
-  loginData: PropTypes.object,
-};
-
-export default function Home({ publicRuntimeConfig }) {
-  const loginData = useLogin('http://localhost:3001/auth/api/v1/user-info');
-
+export default function Home({ publicRuntimeConfig, loginData }) {
   return (
     <div className={styles.container}>
       <Head>
@@ -124,6 +71,7 @@ export default function Home({ publicRuntimeConfig }) {
 
 Home.propTypes = {
   publicRuntimeConfig: PropTypes.object,
+  loginData: PropTypes.object.isRequired,
 };
 
 export async function getServerSideProps() {
